refactor: setting 디자인 컴포넌트 공통 컴포넌트로 변경
